pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepsurvey"
version = "0.1.0"
description = "深入调研并且完成综述"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "criscuolosubidu"}
]
dependencies = [
    "arxiv>=2.2.0",
    "crawl4ai>=0.6.3",
    "gevent>=25.5.1",
    "google>=3.0.0",
    "google-genai>=1.15.0",
    "langchain==0.2.5",
    "langchain-community==0.2.5",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "responses>=0.25.7",
    "tabulate>=0.9.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}
include-package-data = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.pytest.ini_options]
markers = [
    "integration: 标记集成测试",
]
